Update 4.0.1.1
Minor update
long int nomor --> long long int nomor

+++ Chalenge 1/header.h
@@ -48,7 +48,7 @@ float hargaIkan(char *jenisIkan, float berat, int kedalaman);
 void generateOrderID(char * orderID);
 void idAuth(char *orderID, int *percobaan);
 
-void initializeDataPembeli(string nama, long long int *nomor, string alamat, int *kodepos);
+void initializeDataPembeli(string nama, long int *nomor, string alamat, int *kodepos);
 void initializeIkanData(string jenisIkan, string namaIkan, int *kedalaman, float *berat, float *harga, float *diskon, float *ongkir, float *HargaTotal);
