+++ Chalenge 1/main.c
@@ -14,7 +14,7 @@ int main(){
     int hari, bulan, tahun;
     int pilihan;
     int kodepos;
-    long int nomor;
+    long long int nomor;
     bool menu1,menu2,menu3,menu4,menu5,menu6,menu8;
